+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.2.1] 2025-07-28
+
+### Changed
+
+- ML_datasetmaster: `SimpleDatasetMaker` now has a default ID attribute set to None and can be manually set to any string if needed.
+- ML_callbacks: Enhanced docstrings.
+- ML_trainer: Enhanced docstrings.
+
+### Fixed
+
+- ML_optimization: `create_pytorch_problem()` the inner fitness function will correctly handle continuous and binary values.
+
 ## [5.2.0] 2025-07-28
 
 ### Added